oscar_stripe/requirements.txt
django==1.9
django-oscar==1.4
stripe==1.49.0